requirements.txt
PyGithub
caldav
pyyaml
icalendar
